refactor(architecture): audit directory-structure.md compliance
Architecture Compliance Audit - Document 1 of 23

Code Changes:
- Move ProviderError hierarchy to src/domain/errors/ (fix architecture violation)
- Reorganize src/presentation/api/ → src/presentation/routers/api/
- Rename session_cache.py → session_cache_protocol.py
- Rename session_enricher.py → session_enricher_protocol.py
- Rename common_errors.py → common_error.py
- Remove empty src/application/event_handlers/ directory
- Update all imports across source and test files

Documentation Changes:
- Update Mermaid architecture diagram (EventHandlers in Infrastructure)
- Update protocol list (12 → 27+ protocols)
- Update infrastructure directories (5 → 15 directories)
- Add src/schemas/ section
- Fix test structure examples (flat naming pattern)
- Update domain errors list (2 → 8 errors)

Closes: F6.4 Document 1
