+++ src/application/commands/handlers/revoke_session_handler.py
@@ -22,7 +22,7 @@ from src.application.commands.session_commands import RevokeSession
 from src.core.result import Failure, Result, Success
 from src.domain.events.session_events import SessionRevokedEvent
 from src.domain.protocols.event_bus_protocol import EventBusProtocol
-from src.domain.protocols.session_cache import SessionCache
+from src.domain.protocols.session_cache_protocol import SessionCache
 from src.domain.protocols.session_repository import SessionRepository
