+++ src/core/errors/common_error.py
