+++ src/domain/protocols/session_cache_protocol.py
