+++ src/domain/protocols/session_enricher_protocol.py
