+++ src/infrastructure/errors/__init__.py
@@ -1,10 +1,15 @@
 """Infrastructure errors package.
 
-Exports all infrastructure-level error classes for convenient importing.
+Exports infrastructure-level error classes for convenient importing.
 
 Usage:
     from src.infrastructure.errors import DatabaseError, CacheError
-    from src.infrastructure.errors import ProviderAuthenticationError
+
+Note:
+    Provider errors (ProviderError, ProviderAuthenticationError, etc.) are
+    defined in src.domain.errors because they are part of the ProviderProtocol
+    contract. Import them from domain:
+        from src.domain.errors import ProviderError, ProviderAuthenticationError
 """
 
 from src.infrastructure.errors.infrastructure_error import (
@@ -12,11 +17,6 @@ from src.infrastructure.errors.infrastructure_error import (
     DatabaseError,
     ExternalServiceError,
     InfrastructureError,
-    ProviderAuthenticationError,
-    ProviderError,
-    ProviderInvalidResponseError,
-    ProviderRateLimitError,
-    ProviderUnavailableError,
 )
 
 __all__ = [
@@ -24,10 +24,4 @@ __all__ = [
     "DatabaseError",
     "CacheError",
     "ExternalServiceError",
-    # Provider errors
-    "ProviderError",
-    "ProviderAuthenticationError",
-    "ProviderUnavailableError",
-    "ProviderRateLimitError",
-    "ProviderInvalidResponseError",
 ]
