+++ src/presentation/routers/api/middleware/trace_middleware.py
