+++ src/presentation/routers/api/v1/email_verifications.py
@@ -13,7 +13,7 @@ from src.application.commands.auth_commands import VerifyEmail
 from src.application.commands.handlers.verify_email_handler import VerifyEmailHandler
 from src.core.container import get_verify_email_handler
 from src.core.result import Failure, Success
-from src.presentation.api.middleware.trace_middleware import get_trace_id
+from src.presentation.routers.api.middleware.trace_middleware import get_trace_id
 from src.schemas.auth_schemas import (
     AuthErrorResponse,
     EmailVerificationCreateRequest,
