+++ src/presentation/routers/api/v1/password_resets.py
@@ -25,7 +25,7 @@ from src.core.container import (
     get_request_password_reset_handler,
 )
 from src.core.result import Failure, Success
-from src.presentation.api.middleware.trace_middleware import get_trace_id
+from src.presentation.routers.api.middleware.trace_middleware import get_trace_id
 from src.schemas.auth_schemas import (
     AuthErrorResponse,
     PasswordResetCreateRequest,
