+++ src/presentation/routers/api/v1/sessions.py
@@ -55,7 +55,7 @@ from src.core.container import (
     get_revoke_session_handler,
 )
 from src.core.result import Failure, Success
-from src.presentation.api.middleware.trace_middleware import get_trace_id
+from src.presentation.routers.api.middleware.trace_middleware import get_trace_id
 from src.schemas.auth_schemas import (
     AuthErrorResponse,
     SessionCreateRequest,
