+++ src/presentation/routers/api/v1/transactions.py
@@ -44,9 +44,9 @@ from src.core.container import (
     get_sync_transactions_handler,
 )
 from src.core.result import Failure
-from src.presentation.api.middleware.auth_dependencies import AuthenticatedUser
-from src.presentation.api.middleware.trace_middleware import get_trace_id
-from src.presentation.api.v1.errors import ErrorResponseBuilder
+from src.presentation.routers.api.middleware.auth_dependencies import AuthenticatedUser
+from src.presentation.routers.api.middleware.trace_middleware import get_trace_id
+from src.presentation.routers.api.v1.errors import ErrorResponseBuilder
 from src.schemas.transaction_schemas import (
     SyncTransactionsRequest,
     SyncTransactionsResponse,
