+++ tests/api/test_admin_rotation_api.py
@@ -23,7 +23,7 @@ from fastapi import FastAPI, Request, status
 from fastapi.responses import JSONResponse
 from fastapi.testclient import TestClient
 
-from src.presentation.api.middleware.trace_middleware import TraceMiddleware
+from src.presentation.routers.api.middleware.trace_middleware import TraceMiddleware
 from src.schemas.auth_schemas import AuthErrorResponse
 from src.schemas.rotation_schemas import (
     GlobalRotationResponse,
