+++ tests/api/test_providers_endpoints.py
@@ -179,7 +179,9 @@ def mock_connection(mock_connection_id, mock_user_id, mock_provider_id):
 @pytest.fixture(autouse=True)
 def override_auth(mock_user_id):
     """Override authentication for all tests."""
-    from src.presentation.api.middleware.auth_dependencies import get_current_user
+    from src.presentation.routers.api.middleware.auth_dependencies import (
+        get_current_user,
+    )
 
     mock_user = MockCurrentUser(user_id=mock_user_id)
