+++ tests/api/test_transactions_endpoints.py
@@ -234,7 +234,9 @@ def mock_transaction(mock_transaction_id, mock_account_id):
 @pytest.fixture(autouse=True)
 def override_auth(mock_user_id):
     """Override authentication for all tests."""
-    from src.presentation.api.middleware.auth_dependencies import get_current_user
+    from src.presentation.routers.api.middleware.auth_dependencies import (
+        get_current_user,
+    )
 
     mock_user = MockCurrentUser(user_id=mock_user_id)
