+++ tests/integration/test_schwab_transactions_api.py
@@ -22,7 +22,7 @@ import pytest
 from pytest_httpx import HTTPXMock
 
 from src.core.result import Failure, Success
-from src.infrastructure.errors import (
+from src.domain.errors import (
     ProviderAuthenticationError,
     ProviderError,
     ProviderRateLimitError,
